refactor! commands dir
