+++ packages/cloudflare/src/cli/commands/preview.ts
@@ -1,8 +1,8 @@
 import { BuildOptions } from "@opennextjs/aws/build/helper.js";
 import { OpenNextConfig } from "@opennextjs/aws/types/open-next.js";
 
-import { populateCache } from "../populate-cache/populate-cache.js";
 import { getWranglerEnvironmentFlag, runWrangler } from "../utils/run-wrangler.js";
+import { populateCache } from "./populate-cache.js";
 
 export async function preview(
   options: BuildOptions,
